chore: style theme colors.ts, spacing.ts, typography.ts로 분리
