+++ src/styles/emotion.d.ts
@@ -1,5 +1,5 @@
 import "@emotion/react";
-import type { AppTheme } from "./theme";
+import type { AppTheme } from "@/styles/theme/index";
 
 declare module "@emotion/react" {
   export interface Theme extends AppTheme {}
